requirements.txt
# Core Web Framework
Flask==2.2.5
Flask-RESTful==0.3.10
Werkzeug==2.2.3
gunicorn==21.2.0
Jinja2==3.1.2

# Authentication & Security
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
flask-cors==4.0.0
PyJWT==2.8.0
cryptography==41.0.3

# Database & ORM
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.11.3
redis==4.6.0
Flask-Migrate

# AI & ML
langchain==0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.8
langchain-openai>=0.0.2
sentence-transformers==2.2.2
huggingface-hub==0.17.3
pydantic==2.3.0
pandas==2.1.0
numpy==1.25.2

# Vector Databases
chromadb==0.4.13
weaviate-client==3.25.1

# Document Processing
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
openpyxl==3.1.2
python-pptx==0.6.21
beautifulsoup4==4.12.2
requests==2.31.0
html2text==2020.1.16
lxml==4.9.3
pdfplumber
pillow

# Asynchronous Tasks
celery==5.3.4
flower==2.0.1

# File Storage
boto3==1.28.40
minio==7.1.16
python-magic==0.4.27

# Visualization
plotly==5.16.1
matplotlib==3.7.2
seaborn==0.12.2

# UI Components
tailwindcss==0.0.1
Flask-Assets==2.0
jsmin==3.0.1
cssmin==0.2.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
loguru==0.7.0
validators==0.22.0


#Azure
azure-storage-blob